+++ server/src/klite/Browser.kt
@@ -8,8 +8,10 @@ fun detectBrowser(userAgent: String?) = userAgent?.run {
     return if (p >= 0) substring(lastIndexOf(' ', p) + 1, indexOf(' ', startIndex = p + 1).takeIf { it > 0 } ?: length) else null
   }
 
-  (if (contains("Mobile")) "Mobile/" else "") + (
+  val bot = split(' ', ';').find { it.contains("bot") } ?: ""
+  val primary = (if (contains("Mobile")) "Mobile/" else "") + (
     detect("Edg") ?: detect("Chrome") ?: detect("Firefox") ?: detect("Trident") ?: detect("MSIE") ?: detect("iOS") ?:
     if (contains("Safari")) detect("Version")?.replace("Version", "Safari")
-    else detect("AppleWebKit") ?: split(' ', ';').find { it.contains("bot") } ?: this)?.replace(".0.0.0", "")
+    else detect("AppleWebKit") ?: (if (bot.isEmpty()) this else ""))?.replace(".0.0.0", "")
+  listOfNotNull(bot, primary).filter { it.isNotEmpty() }.joinToString("/")
 }